+++ src/app/components/connection/connection.component.ts
@@ -1,11 +1,12 @@
 import { Component, OnInit } from '@angular/core';
-import { Connection } from 'dist/sitmun-frontend-core/connection/connection.model';
-import { ConnectionService } from 'dist/sitmun-frontend-core/';
+import { ConnectionService, Connection } from '@sitmun/frontend-core';
 import { UtilsService } from '../../services/utils.service';
 
 import { Router } from '@angular/router';
 import { environment } from 'src/environments/environment';
 import { Subject } from 'rxjs';
+import { MatDialog } from '@angular/material/dialog';
+import { DialogMessageComponent } from 'dist/sitmun-frontend-gui/';
 
 @Component({
   selector: 'app-connection',
@@ -18,7 +19,8 @@ export class ConnectionComponent implements OnInit {
   themeGrid: any = environment.agGridTheme;
   columnDefs: any[];
 
-  constructor(public connectionService: ConnectionService,
+  constructor(public dialog: MatDialog,
+    public connectionService: ConnectionService,
     private utils: UtilsService,
     private router: Router,
   ) {
@@ -26,29 +28,17 @@ export class ConnectionComponent implements OnInit {
   }
 
   ngOnInit() {
+
+    var columnEditBtn=environment.editBtnColumnDef;
+    columnEditBtn['cellRendererParams']= {
+      clicked: this.newData.bind(this)
+    }
+
+
     this.columnDefs = [
-      {
-        headerName: '',
-        checkboxSelection: true,
-        headerCheckboxSelection: true,
-        editable: false,
-        filter: false,
-        width: 40,
-        lockPosition: true,
-      },
-      {
-        headerName: '',
-        field: 'id',
-        editable: false,
-        filter: false,
-        width: 48,
-        lockPosition: true,
-        cellRenderer: 'btnEditRendererComponent',
-        cellRendererParams: {
-          clicked: this.newData.bind(this)
-        },
-      },
-      { headerName: 'Id', field: 'id', editable: false },
+      environment.selCheckboxColumnDef,
+     columnEditBtn,
+      { headerName: 'Id', field: 'id', editable: false},
       { headerName: this.utils.getTranslate('connectionEntity.name'), field: 'name' },
       { headerName: this.utils.getTranslate('connectionEntity.user'), field: 'user' },
       { headerName: this.utils.getTranslate('connectionEntity.driver'), field: 'driver' },
@@ -88,14 +78,23 @@ export class ConnectionComponent implements OnInit {
   }
 
   removeData(data: Connection[]) {
-    const promises: Promise<any>[] = [];
-    data.forEach(connection => {
-      promises.push(new Promise((resolve, reject) => {​​​​​​​ this.connectionService.delete(connection).toPromise().then((resp) =>{​​​​​​​resolve()}​​​​​​​)}​​​​​​​));
-      Promise.all(promises).then(() => {
-        this.dataUpdatedEvent.next(true);
-      });
-    });
 
+    const dialogRef = this.dialog.open(DialogMessageComponent);
+    dialogRef.componentInstance.title=this.utils.getTranslate("caution");
+    dialogRef.componentInstance.message=this.utils.getTranslate("removeMessage");
+    dialogRef.afterClosed().subscribe(result => {
+      if(result){
+        if(result.event==='Accept') {  
+          const promises: Promise<any>[] = [];
+          data.forEach(connection => {
+            promises.push(new Promise((resolve, reject) => {​​​​​​​ this.connectionService.delete(connection).toPromise().then((resp) =>{​​​​​​​resolve()}​​​​​​​)}​​​​​​​));
+            Promise.all(promises).then(() => {
+              this.dataUpdatedEvent.next(true);
+            });
+          });
+       }
+      }
+    });
   }
 
 }
